app/requirements.txt
Flask==3.0.0
boto3==1.34.153
PyYAML==6.0.2
gunicorn==22.0.0